hagrid/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.3
fastapi==0.73.0
grpcio==1.43.0
grpcio-tools==1.43.0
h11==0.13.0
httptools==0.3.0
idna==3.3
protobuf==3.19.4
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.4
uvloop==0.16.0
watchgod==0.7
websockets==10.1
